Refactor : split Collection into Traits
